Move module resolver
