+++ lib/leaf.js
@@ -1,7 +1,7 @@
 module.exports = Leaf;
 
 var HAS_DEFINE = /define\(/;
-var IS_RELATIVE = /^(\.|\.\.)\//;
+var IS_RELATIVE = /^\.\//;
 
 var esprima = require('esprima');
 var Scope = require('./scope');
@@ -12,25 +12,9 @@ var VariableDeclarator = require('./visitors/variable-declarator');
 var DefineCallExpression = require('./visitors/define-call-expression');
 var cloneDeep = require('lodash-node/modern/objects/cloneDeep');
 var path = require('path');
+var resolveModule = require('./helpers/resolve-module');
 
 module.exports = Leaf;
-
-var resolveModule = function ( child, name ) {
-  if (child.charAt(0) !== '.') { return child; }
-  var parts = child.split('/');
-  var nameParts = name.split('/');
-  var parentBase = nameParts.length > 1 ? nameParts.slice(0, -1) : [ nameParts[ 0 ] ];
-
-  for (var i = 0, l = parts.length; i < l; i++) {
-    var part = parts[i];
-
-    if (part === '..') { parentBase.pop(); }
-    else if (part === '.') { continue; }
-    else { parentBase.push(part); }
-  }
-
-  return parentBase.join('/');
-};
 
 function Leaf(path, source, ast) {
   this.path = path;
